Update show_alloc_mem();
Changed the whole function, as I result that I was first trying to
create it using my own printf(); and it is not allowed, so I recreated
it with some new functions to handlle conversion.

Created number_to_hex() that will take an uint64_t as a parameter that
is the address of the chunk and convert it to hexadecimal numbers.
Also created a function put_number();