Write a function to show memory history
A very simple function that will try to open a file created before by
malloc() in the form of .<pid_id>_mem_historic, on successful the
function will read one character at a time until reaches "\n" and save
it to an array, once "\n" is found print the line, if EOF stop, we are
done.

It would be much better to read more than on char, as I'm coding
malloc() I cannot use it, i.e call my get_next_line(); needs malloc();
and as I have to check for "\n" line and EOF the best way I found was
reading one character at a time.